Sparse vector search benchmarks
