+++ lib/segment/src/fixtures/mod.rs
@@ -2,3 +2,4 @@ pub mod index_fixtures;
 pub mod payload_context_fixture;
 pub mod payload_fixtures;
 pub mod segment_fixtures;
+pub mod sparse_fixtures;
